% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A1}
\alias{df_A1}
\title{Simulated Clinical Data for A1 Unsupervised Clustering}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A1
}
\description{
A simulated clinical dataset used to demonstrate the functionality of the A1_unsupervised_cluster algorithm in the AFKAR package.
}
\keyword{datasets}
