% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A2}
\alias{df_A2}
\title{Simulated Clinical Data for A2 Anomaly Detection}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A2
}
\description{
A simulated dataset used to demonstrate the A2_unsupervised_anomaly detection algorithm.
}
\keyword{datasets}
