% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A3}
\alias{df_A3}
\title{Simulated Clinical Data for A3 Dimensionality Reduction}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A3
}
\description{
A simulated dataset used to demonstrate the A3_unsupervised_reduction algorithm.
}
\keyword{datasets}
