% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A4}
\alias{df_A4}
\title{Simulated Clinical Data for A4 Association Rule Mining}
\format{
A data.frame containing binary clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A4
}
\description{
A simulated binary dataset used to demonstrate the A4_unsupervised_association algorithm.
}
\keyword{datasets}
