% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A6}
\alias{df_A6}
\title{Simulated Clinical Data for A6 Network Community Detection}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A6
}
\description{
A simulated dataset used to demonstrate the A6_unsupervised_network syndemic architecture mapping.
}
\keyword{datasets}
