% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_A7}
\alias{df_A7}
\title{Simulated Clinical Data for A7 Time-Series Trajectory Clustering}
\format{
A data.frame containing sequential clinical measurements.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_A7
}
\description{
A simulated time-series dataset used to demonstrate the A7_unsupervised_timeseries clustering algorithm.
}
\keyword{datasets}
