% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_B1}
\alias{df_B1}
\title{Simulated Clinical Data for B1 Diagnostic Screening Classifier}
\format{
A data.frame containing clinical predictors and a binary target variable.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_B1
}
\description{
A simulated dataset used to demonstrate the B1_supervised_screening logistic regression model.
}
\keyword{datasets}
