% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_B2}
\alias{df_B2}
\title{Simulated Clinical Data for B2 Emergency Triage Predictor}
\format{
A data.frame containing vital signs and an ICU admission target.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_B2
}
\description{
A simulated dataset used to demonstrate the B2_supervised_triage Early Warning System model.
}
\keyword{datasets}
