% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_B4}
\alias{df_B4}
\title{Simulated Clinical Data for B4 Supervised Learning}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_B4
}
\description{
A simulated dataset used to demonstrate the B4 supervised learning algorithm.
}
\keyword{datasets}
