% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_B9}
\alias{df_B9}
\title{Simulated Clinical Data for B9 Supervised Learning}
\format{
A data.frame containing simulated clinical features.
}
\source{
Simulated clinical data for AFKAR package.
}
\usage{
df_B9
}
\description{
A simulated dataset used to demonstrate the B9 supervised learning algorithm.
}
\keyword{datasets}
